[fixup] Address review comments (to squash with commit '[FLINK-19162][connectors] Add 'recycle()' to the RecordsWithSplitIds to support reuse of heavy objects.')
